@@ -16,10 +16,11 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 1.0.0+1
+version: 1.0.0+6
 
 environment:
-  sdk: '>=2.19.6 <3.0.0'
+  sdk: '>=3.0.3 <4.0.0'
+
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -30,43 +31,46 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  flutter_svg: ^2.0.5
-  google_fonts: ^4.0.4
-
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  font_awesome_flutter: ^10.4.0
-  firebase_core: ^2.13.1
-  firebase_auth: ^4.6.1
-  googleapis_auth: ^1.4.0
-  url_launcher: ^6.1.11
-  file_picker: ^5.3.0
-  google_sign_in: ^6.1.1
-  http: ^0.13.6
+
+  cloud_firestore: ^4.8.2
+  google_fonts: ^4.0.4
   flutter_web_auth: ^0.5.0
-  googleapis: ^11.1.0
   web_socket_channel: ^2.4.0
-  extension_google_sign_in_as_googleapis_auth: ^2.0.9
-  firebase_storage: ^11.2.2
-  cloud_firestore: ^4.7.1
-  flutter_staggered_grid_view: ^0.6.2
-  cached_network_image: ^3.2.3
-  grouped_list: ^5.1.2
-  firebase_ui_auth: ^1.4.2
-  firebase_ui_storage: ^0.1.0-dev.5
-  firebase_ui_firestore: 1.5.2
-  flutter_blurhash: ^0.7.0
-  intl: ^0.18.0
-  flutter_phone_direct_caller: ^2.1.1
-  flutter_scrolling_fab_animated: ^0.1.4
-  firebase_database: ^10.2.2
-  flutter_cache_manager: ^3.3.0
-  image_picker: ^0.8.8
+  firebase_storage: ^11.2.4
+  firebase_core: ^2.14.0
   photo_view: ^0.14.0
-  flutter_platform_widgets: ^3.3.5
+  flutter_launcher_icons: ^0.12.0
+  flutter_riverpod: ^2.3.6
   uuid: ^3.0.7
+  flutter_platform_widgets: ^3.3.5
+  image_picker: ^1.0.0
+  flutter_cache_manager: ^3.3.1
+  flutter_scrolling_fab_animated: ^0.1.4
+  flutter_phone_direct_caller: ^2.1.1
+  intl: ^0.18.0
+  flutter_blurhash: ^0.7.0
+  grouped_list: ^5.1.2
+  cached_network_image: ^3.2.3
+  firebase_auth: ^4.6.3
+  url_launcher: ^6.1.11
+  flutter_staggered_grid_view: ^0.6.2
+  flutter_svg: ^2.0.7
+  font_awesome_flutter: ^10.5.0
+  video_player: ^2.7.0
+  pinput: ^2.2.31
+  firebase_ui_firestore: ^1.5.4
+  image: ^4.0.17
+  blurhash_dart: ^1.2.1
+  share_plus: ^7.0.2
+  pull_down_button: ^0.9.1
+  modal_bottom_sheet: ^2.1.2
+  animations: ^2.0.7
+
+
 
 dev_dependencies:
   flutter_test:
