@@ -1,9 +1,16 @@
-import 'dart:ui';
-
 import 'package:flutter/material.dart';
 
-const backgroundColor = Color.fromRGBO(60, 58, 59, 1);
+const background = Color.fromRGBO(66, 66, 66, 1);
+const secondaryColor = Color.fromRGBO(251, 174, 27,1);
+const greenShade600 = Color.fromRGBO(90, 100, 63, 1);
 const primaryColor = Color.fromRGBO(84, 154, 82, 1);
-const secondaryColor = Color.fromRGBO(251, 174, 27, 1);
-const banner = Color.fromRGBO(167, 93, 53, 1);
+const brownShade800 = Color.fromRGBO(48, 32, 16, 1);
+const brownShade600 = Color.fromRGBO(72, 45, 18, 1);
+const brownShade300 = Color.fromRGBO(167, 93, 53, 1);
+const greyShade700 = Color.fromRGBO(60, 58, 59, 1);
+const greyShade600 = Color.fromRGBO(74, 73, 72, 1);
+const greyShade400 = Color.fromRGBO(122, 121, 121, 1);
+const greyShade200 = Color.fromRGBO(190, 190, 190, 1);
+const greyShade100 = Color.fromRGBO(222, 222, 222, 1);
+const greyShade10 = Color.fromRGBO(235, 235, 235, 1);
 const white = Colors.white;
