@@ -1,26 +1,115 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:google_fonts/google_fonts.dart';
+import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
+import 'package:regiment8112_project/data/plattons.dart';
+import 'package:regiment8112_project/providers/user_provider.dart';
+import 'package:regiment8112_project/providers/validatorProvider.dart';
+import 'package:regiment8112_project/screens/main_screen.dart';
+import 'package:regiment8112_project/services/user_service.dart';
 import 'package:regiment8112_project/utils/colors.dart';
+import 'package:regiment8112_project/utils/validators.dart';
 import 'package:regiment8112_project/widgets/custom_text.dart';
+import 'package:regiment8112_project/widgets/custom_text_field.dart';
 
-class AddContact extends StatelessWidget {
+class AddContact extends ConsumerStatefulWidget {
   const AddContact({super.key});
 
   @override
-  Widget build(BuildContext context) {
-    TextEditingController controller = TextEditingController();
+  ConsumerState<AddContact> createState() => _AddContactState();
+}
 
+final UserService _service = UserService();
+
+class _AddContactState extends ConsumerState<AddContact> {
+  String? _selectedVal = 'מחלקה 1';
+  int _selectedValue = -1;
+  bool isError = false;
+
+  final _formField = GlobalKey<FormState>();
+  final _nameController = TextEditingController();
+  final _lastNameController = TextEditingController();
+  final _phoneController = TextEditingController();
+  final _cityController = TextEditingController();
+
+  @override
+  void dispose() {
+    super.dispose();
+    _nameController.dispose();
+    _lastNameController.dispose();
+    _cityController.dispose();
+    _phoneController.dispose();
+  }
+
+  void createUserCupertino(TargetPlatform platform) {
+    if (_formField.currentState!.validate()) {
+      ref.watch(userProvider.notifier).getUser().then((value) {
+        if (value == true) {
+        } else {
+          _service.addContact(
+              _nameController.text.trim(),
+              _lastNameController.text.trim(),
+              _phoneController.text.trim(),
+              _cityController.text.trim(),
+              platoon[_selectedValue]);
+          Navigator.pushReplacement(
+              context,
+              platform == TargetPlatform.iOS
+                  ? CupertinoPageRoute(
+                      builder: (context) =>
+                          CupertinoScaffold(body: const MainScreen()))
+                  : MaterialWithModalsPageRoute(
+                      builder: (context) => const MainScreen()));
+        }
+      });
+    }
+  }
+
+  void createUserMaterial(TargetPlatform platform) {
+    if (_formField.currentState!.validate()) {
+      ref.watch(userProvider.notifier).getUser().then((value) {
+        if (value == true) {
+        } else {
+          _service.addContact(
+              _nameController.text.trim(),
+              _lastNameController.text.trim(),
+              _phoneController.text.trim(),
+              _cityController.text.trim(),
+              _selectedVal!);
+          Navigator.pushReplacement(context,
+              platform == TargetPlatform.iOS
+                  ? CupertinoPageRoute(
+                  builder: (context) =>
+                      CupertinoScaffold(body: const MainScreen()))
+                  : MaterialWithModalsPageRoute(
+                  builder: (context) => const MainScreen()));
+        }
+      });
+    }
+  }
+
+  Widget container(List<Widget> child) {
     final size = MediaQuery.of(context).size;
+    final isDark = Theme.of(context).brightness == Brightness.dark;
+    final colorScheme = Theme.of(context).colorScheme;
+    final platform = Theme.of(context).platform;
 
-    return PlatformApp(
-      material: (_, __) => MaterialAppData(
-        debugShowCheckedModeBanner: false,
-        home: Scaffold(
-          body: Container(
-            height: double.infinity,
-            color: backgroundColor,
+    return Material(
+      child: Form(
+        key: _formField,
+        child: Container(
+          height: size.height,
+          decoration: BoxDecoration(
+              image: DecorationImage(
+                  opacity: isDark ? 0.12 : 1,
+                  image: AssetImage("assets/images/Group 126.png"),
+                  fit: BoxFit.cover),
+              color: colorScheme.background),
+          child: Padding(
+            padding:
+                const EdgeInsets.symmetric(horizontal: 30.0, vertical: 50.0),
             child: SingleChildScrollView(
               child: Directionality(
                 textDirection: TextDirection.rtl,
@@ -37,63 +126,48 @@ class AddContact extends StatelessWidget {
                         )
                       ],
                     ),
+                    Text("הירשם על מנת להמשיך",
+                        style: GoogleFonts.rubikDirt(
+                            fontSize: 32, color: colorScheme.primary)),
+                    CustomText(
+                        text:
+                            "בצע רישום חד פעמי ולאחר מכן תוכל להמשיך לאפליקציה",
+                        fontSize: 12),
                     Padding(
-                      padding: const EdgeInsets.all(30.0),
-                      child: SizedBox(
-                        height: size.width,
-                        child: Column(
-                          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                          children: [
-                            TextField(
-                              style: const TextStyle(color: white),
-                              controller: controller,
-                              decoration: const InputDecoration(
-                                  border: OutlineInputBorder(),
-                                  labelText: 'בן אנוש',
-                                  disabledBorder: InputBorder.none,
-                                  filled: true,
-                                  fillColor: white),
-                            ),
-                            TextField(
-                              controller: controller,
-                              decoration: const InputDecoration(
-                                  border: OutlineInputBorder(),
-                                  labelText: 'בן אנוש',
-                                  filled: true,
-                                  fillColor: white),
-                            ),
-                            TextField(
-                              controller: controller,
-                              decoration: const InputDecoration(
-                                  border: OutlineInputBorder(),
-                                  labelText: 'בן אנוש',
-                                  filled: true,
-                                  fillColor: white),
-                            ),
-                            TextField(
-                              controller: controller,
-                              decoration: const InputDecoration(
-                                  border: OutlineInputBorder(),
-                                  filled: true,
-                                  labelText: 'בן אנוש',
-                                  fillColor: white),
-                            ),
-                          ],
-                        ),
+                      padding: const EdgeInsets.only(top: 25.0),
+                      child: Column(
+                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                        children: child,
                       ),
                     ),
-                    SizedBox(
-                      width: size.width / 2,
-                      child: ElevatedButton(
-                          style: ButtonStyle(
-                              backgroundColor:
-                                  MaterialStateProperty.all(primaryColor)),
-                          onPressed: () {},
-                          child: const CustomText(
-                            fontSize: 16,
-                            color: white,
-                            text: "הוסף איש קשר",
-                          )),
+                    Padding(
+                      padding: const EdgeInsets.only(top: 50.0),
+                      child: SizedBox(
+                        height: 56,
+                        width: size.width / 2,
+                        child: Directionality(
+                          textDirection: TextDirection.rtl,
+                          child: ElevatedButton.icon(
+                            icon: Icon(Icons.adaptive.arrow_back),
+                            label: const CustomText(
+                              fontSize: 16,
+                              text: "המשך לאפליקציה",
+                              color: white,
+                            ),
+                            style: ButtonStyle(
+                                backgroundColor: MaterialStateProperty.all(
+                                    isDark ? primaryColor : secondaryColor)),
+                            onPressed: () {
+                              if (Theme.of(context).platform ==
+                                  TargetPlatform.iOS) {
+                                createUserCupertino(platform);
+                              } else {
+                                createUserMaterial(platform);
+                              }
+                            },
+                          ),
+                        ),
+                      ),
                     )
                   ],
                 ),
@@ -102,26 +176,202 @@ class AddContact extends StatelessWidget {
           ),
         ),
       ),
-      cupertino: (_, __) => CupertinoAppData(
-        home: PlatformScaffold(
-          cupertino: (_, __) => CupertinoPageScaffoldData(
-            body: Center(
-              child: Container(
-                color: backgroundColor,
-                child: CupertinoFormSection(
-                  children: [
-                    CupertinoTextField(
-                        textDirection: TextDirection.rtl,
-                        placeholder: "חפש בן אדם",
-                        placeholderStyle:
-                            GoogleFonts.heebo(color: primaryColor))
-                  ],
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final size = MediaQuery.of(context).size;
+    var validateProvider = ref.watch(validatorProvider.notifier);
+
+    return PlatformScaffold(
+      material: (_, __) => MaterialScaffoldData(
+          body: container([
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            CustomTextField(
+              controller: _nameController,
+              text: "שם פרטי",
+              maxLength: null,
+              width: size.width * 0.42,
+              textInputAction: TextInputAction.next,
+              validator: (value) {
+                return validateProvider.validator(
+                    value!,
+                    "שם אינו יכול להיות ריק!",
+                    "השם שהוזן אינו תקין!",
+                    nameValidator);
+              },
+            ),
+            CustomTextField(
+                controller: _lastNameController,
+                text: "שם משפחה",
+                maxLength: null,
+                width: size.width * 0.42,
+                textInputAction: TextInputAction.next,
+                validator: (value) {
+                  return validateProvider.validator(
+                      value!,
+                      "שם משפחה אינו יכול להיות ריק!",
+                      "שם המשפחה שהוזן אינו תקין!",
+                      nameValidator);
+                }),
+          ],
+        ),
+        CustomTextField(
+          controller: _cityController,
+          text: "עיר מגורים",
+          maxLength: null,
+          textInputAction: TextInputAction.next,
+        ),
+        CustomTextField(
+          controller: _phoneController,
+          text: "מספר טלפון",
+          textInputAction: TextInputAction.done,
+          type: TextInputType.phone,
+          validator: (value) {
+            return validateProvider.validator(value!, "אנא מלא מספר טלפון",
+                "נא הקש מספר טלפון תקין!", phoneValidator);
+          },
+        ),
+        SizedBox(
+          width: double.infinity,
+          child: pickerMaterial(),
+        )
+      ])),
+      cupertino: (_, __) => CupertinoPageScaffoldData(
+        body: container(
+          [
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                CustomTextField(
+                    textInputAction: TextInputAction.next,
+                    validator: (value) {
+                      return validateProvider.validator(
+                          value!,
+                          "שם אינו יכול להיות ריק!",
+                          "שם שהוזן אינו תקין!",
+                          nameValidator);
+                    },
+                    controller: _nameController,
+                    text: "שם פרטי",
+                    width: size.width / 2.4),
+                CustomTextField(
+                    validator: (value) {
+                      return validateProvider.validator(
+                          value!,
+                          "שם משפחה אינו יכול להיות ריק!",
+                          "שם המשפחה שהוזן אינו תקין!",
+                          nameValidator);
+                    },
+                    controller: _lastNameController,
+                    textInputAction: TextInputAction.next,
+                    text: "שם משפחה",
+                    width: size.width / 2.4),
+              ],
+            ),
+            CustomTextField(
+              controller: _phoneController,
+              textInputAction: TextInputAction.next,
+              text: "מספר טלפון",
+              type: TextInputType.phone,
+              validator: (value) {
+                return validateProvider.validator(value!, "אנא מלא מספר טלפון",
+                    "נא הקש מספר טלפון תקין!", phoneValidator);
+              },
+            ),
+            CustomTextField(controller: _cityController, text: "עיר מגורים"),
+            CustomTextField(
+              controller: null,
+              textInputAction: TextInputAction.next,
+              text:
+                  _selectedValue == -1 ? 'בחר מחלקה' : platoon[_selectedValue],
+              readOnly: true,
+              prefix: GestureDetector(
+                child: const Icon(
+                  CupertinoIcons.chevron_down,
+                  color: primaryColor,
                 ),
+                onTap: () {
+                  showCupertinoModalPopup(
+                    context: context,
+                    builder: (context) => Container(child: pickerCupertino()),
+                  );
+                },
               ),
             ),
-          ),
+          ],
         ),
       ),
+    );
+  }
+
+  Widget pickerMaterial() {
+    bool isDark = Theme.of(context).brightness == Brightness.dark;
+    return DropdownButtonFormField<String>(
+        value: _selectedVal,
+        decoration: InputDecoration(
+            floatingLabelBehavior: FloatingLabelBehavior.never,
+            focusedBorder: const OutlineInputBorder(
+              borderRadius: BorderRadius.all(Radius.circular(10.0)),
+              borderSide: BorderSide(color: primaryColor, width: 2.0),
+            ),
+            border: const OutlineInputBorder(
+              borderSide: BorderSide.none,
+              borderRadius: BorderRadius.all(Radius.circular(10)),
+            ),
+            labelText: "בחר מחלקה",
+            filled: true,
+            fillColor: isDark ? white.withOpacity(0.1) : greyShade100),
+        hint: const CustomText(text: "בחר מחלקה", fontSize: 16),
+        alignment: Alignment.center,
+        items: platoon
+            .map((item) => DropdownMenuItem<String>(
+                  value: item,
+                  child: CustomText(text: item, fontSize: 16),
+                ))
+            .toList(),
+        onChanged: (value) {
+          setState(() {
+            _selectedVal = value;
+          });
+        });
+  }
+
+  Widget pickerCupertino() {
+    return CupertinoActionSheet(
+      cancelButton: CupertinoButton(
+        child: const Text("ביטול"),
+        onPressed: () {
+          setState(() {
+            _selectedValue = -1;
+          });
+          Navigator.pop(context);
+        },
+      ),
+      actions: [
+        SizedBox(
+          height: 300,
+          child: CupertinoPicker(
+            magnification: 1.2,
+            scrollController:
+                FixedExtentScrollController(initialItem: _selectedValue),
+            looping: true,
+            itemExtent: 50,
+            onSelectedItemChanged: (int value) {
+              setState(() {
+                platoon[value];
+                _selectedValue = value;
+              });
+            },
+            children: [
+              for (String name in platoon) Center(child: Text(name)),
+            ],
+          ),
+        )
+      ],
     );
   }
 }
