@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-
 import '../utils/colors.dart';
 import 'custom_text.dart';
 
@@ -11,22 +10,26 @@ class Bubble extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      mainAxisAlignment: MainAxisAlignment.end,
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        CustomText(fontSize: 12,color:  const Color.fromRGBO(190, 190, 190, 1),text: date),
-        Container(
-          padding: const EdgeInsets.all(20),
-          width: double.infinity,
-          decoration: const BoxDecoration(
-              color: Color.fromRGBO(121, 121, 121, 1),
-              borderRadius: BorderRadius.all(Radius.circular(10))
-          ),
-          child: CustomText(fontSize: 16, color: white ,text: text),
-        )
-      ],
+    final colorScheme = Theme.of(context).colorScheme;
+    final isDark = Theme.of(context).brightness == Brightness.dark;
+    return Container(
+      width: double.infinity,
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.end,
+        children: [
+          CustomText(fontSize: 12, color: colorScheme.onBackground, text: date,),
+          Directionality(
+            textDirection: TextDirection.rtl,
+            child: Container(
+              padding: const EdgeInsets.all(20),
+              decoration: BoxDecoration(
+                  color: isDark ? greyShade400 : greyShade100,
+                  borderRadius: BorderRadius.all(Radius.circular(10))),
+              child: CustomText(fontSize: 16, color: colorScheme.onBackground, text: text,textAlign: TextAlign.start),
+            ),
+          )
+        ],
+      ),
     );
   }
 }
