@@ -1,36 +1,39 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
-import 'package:regiment8112_project/utils/colors.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:regiment8112_project/providers/search_provider.dart';
 import 'package:regiment8112_project/widgets/next_summon.dart';
 import 'package:regiment8112_project/widgets/search_bar.dart';
 import 'custom_text.dart';
 
-class TopSection extends StatefulWidget {
+class TopSection extends ConsumerStatefulWidget {
   const TopSection(this.currentTab, this.updateActiveTab, {super.key});
 
   final String currentTab;
-
   final Function(String value) updateActiveTab;
 
   @override
-  State<TopSection> createState() => _TopSectionState();
+  ConsumerState<TopSection> createState() => _TopSectionState();
 }
 
-class _TopSectionState extends State<TopSection> {
+class _TopSectionState extends ConsumerState<TopSection> {
   var activeTab = 'news';
+  final _searchController = TextEditingController();
 
   @override
-  void initState() {
-    super.initState();
-  }
-
-  void function(BuildContext context, String route, Widget widget) {
-    Navigator.push(context,
-        MaterialPageRoute(maintainState: true, builder: (context) => widget));
+  void dispose() {
+    _searchController.dispose();
+    super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
+    final colorScheme = Theme.of(context).colorScheme;
+    var searchProv = ref.read(searchProvider.notifier);
+    Color isContactList = activeTab == 'contactsList'
+        ? colorScheme.secondary
+        : colorScheme.primary;
+
     return Column(
       children: [
         Row(
@@ -42,32 +45,45 @@ class _TopSectionState extends State<TopSection> {
                 material: (_, __) => MaterialTextButtonData(
                   onPressed: () {
                     setState(() {
+                      searchProv.state = '';
                       activeTab = 'news';
                     });
                     widget.updateActiveTab(activeTab);
                   },
                   child: CustomText(
+                      fontWeight: activeTab == 'news'
+                          ? FontWeight.w600
+                          : FontWeight.w400,
                       fontSize: 16,
-                      color:
-                          activeTab == 'news' ? secondaryColor : primaryColor,
+                      color: activeTab == 'news'
+                          ? colorScheme.secondary
+                          : colorScheme.primary,
                       text: "עדכונים וחדשות"),
                 ),
                 cupertino: (_, __) => CupertinoTextButtonData(onPressed: () {
                   setState(() {
+                    searchProv.state = '';
                     activeTab = 'news';
                   });
                   widget.updateActiveTab(activeTab);
                 }),
                 child: CustomText(
+                    fontWeight:
+                        activeTab == 'news' ? FontWeight.w600 : FontWeight.w400,
                     fontSize: 16,
-                    color: activeTab == 'news' ? secondaryColor : primaryColor,
+                    color: activeTab == 'news'
+                        ? colorScheme.secondary
+                        : colorScheme.primary,
                     text: "עדכונים וחדשות"),
               ),
             ),
-            Image.asset(
-              'assets/svg/logo.png',
-              width: 97,
-              height: 97,
+            Hero(
+              tag: "logo",
+              child: Image.asset(
+                'assets/svg/logo.png',
+                width: 97,
+                height: 97,
+              ),
             ),
             PlatformTextButton(
               material: (_, __) => MaterialTextButtonData(
@@ -78,10 +94,10 @@ class _TopSectionState extends State<TopSection> {
                   widget.updateActiveTab('contactsList');
                 },
                 child: CustomText(
+                    fontWeight:
+                        activeTab != 'news' ? FontWeight.w600 : FontWeight.w400,
                     fontSize: 16,
-                    color: activeTab == 'contactsList'
-                        ? secondaryColor
-                        : primaryColor,
+                    color: isContactList,
                     text: "רשימת קשר"),
                 // )
               ),
@@ -93,16 +109,23 @@ class _TopSectionState extends State<TopSection> {
                   widget.updateActiveTab('contactsList');
                 },
                 child: CustomText(
+                    fontWeight:
+                        activeTab != 'news' ? FontWeight.w600 : FontWeight.w400,
                     fontSize: 16,
-                    color: activeTab == 'contactsList'
-                        ? secondaryColor
-                        : primaryColor,
+                    color: isContactList,
                     text: "רשימת קשר"),
               ),
             )
           ],
         ),
-        activeTab == 'news' ? const NextSummon() : const CustomSearchBar()
+        activeTab == 'news'
+            ? const NextSummon()
+            : CustomSearchBar(
+                controller: _searchController,
+                onChanged: (value) {
+                  searchProv.updateQuery(value);
+                },
+              )
       ],
     );
   }
