@@ -6,17 +6,17 @@
 
 #include "generated_plugin_registrant.h"
 
-#include <desktop_webview_auth/desktop_webview_auth_plugin.h>
 #include <file_selector_linux/file_selector_plugin.h>
+#include <smart_auth/smart_auth_plugin.h>
 #include <url_launcher_linux/url_launcher_plugin.h>
 
 void fl_register_plugins(FlPluginRegistry* registry) {
-  g_autoptr(FlPluginRegistrar) desktop_webview_auth_registrar =
-      fl_plugin_registry_get_registrar_for_plugin(registry, "DesktopWebviewAuthPlugin");
-  desktop_webview_auth_plugin_register_with_registrar(desktop_webview_auth_registrar);
   g_autoptr(FlPluginRegistrar) file_selector_linux_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "FileSelectorPlugin");
   file_selector_plugin_register_with_registrar(file_selector_linux_registrar);
+  g_autoptr(FlPluginRegistrar) smart_auth_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "SmartAuthPlugin");
+  smart_auth_plugin_register_with_registrar(smart_auth_registrar);
   g_autoptr(FlPluginRegistrar) url_launcher_linux_registrar =
       fl_plugin_registry_get_registrar_for_plugin(registry, "UrlLauncherPlugin");
   url_launcher_plugin_register_with_registrar(url_launcher_linux_registrar);
